Include contents of `mmap.rd` into `README.rdoc`
Currently, the gem's main README, or the docs generated with `rake docs`
don't include a description of the interface, or much documentation.

The most useful documentation is in file `mmap.rd` which is not very
obvious.

This change includes the contents of `mmap.rd` into `README.rdoc`, which
is the main port of call for any user of a gem.